requirements.txt
# Core dependencies
sqlglot>=20.0.0
pyarrow>=14.0.0
pyyaml>=6.0

# Data source connectors
psycopg2-binary>=2.9.0
duckdb>=0.9.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
mypy>=1.7.0
